requirements.txt
pyglet
synthizer
accessible_output2
attrs
